♻️ (Auth Button) - Split AuthButton component into smaller reusable components + Move auth-related UI elements into separate components under AuthButton directory for better maintainability and reusability
