+++ src/components/Auth/AuthButton/ForgotPasswordLink.tsx
@@ -1,0 +1,20 @@
+import { Button } from "@/components/ui/button";
+import React from "react";
+
+interface ForgotPasswordLinkProps {
+  setAuthAction: (action: "login" | "signup" | "forgot") => void;
+}
+
+const ForgotPasswordLink: React.FC<ForgotPasswordLinkProps> = ({
+  setAuthAction,
+}) => {
+  return (
+    <div className="text-sm text-right">
+      <Button variant={"link"} onClick={() => setAuthAction("forgot")}>
+        Mot de passe oublié ?
+      </Button>
+    </div>
+  );
+};
+
+export default ForgotPasswordLink;
